+++ .replit
@@ -1,2 +1,2 @@
 language = "cpp"
-run = "clang++ -std=c++17 -Weverything -Wno-padded -Wno-c++98-compat -Wno-exit-time-destructors -Wno-missing-prototypes main.cpp && ./a.out"
+run = "clang++ -std=c++17 -Weverything -Wno-padded -Wno-c++98-compat -Wno-exit-time-destructors -Wno-missing-prototypes main.cpp AirPods.cpp GiftCard.cpp Headphones.cpp HomeTheater.cpp Television.cpp Wrappers.cpp && ./a.out"
