+++ main.cpp
@@ -36,420 +36,14 @@
          just split it up into source files and provide the appropriate #include directives.
  */
 
-
-#include <iostream>
 #include "LeakedObjectDetector.h"
-/*
- copied UDT 1:
- */
-struct Headphones
-{
-    Headphones();
-    ~Headphones() 
-    {
-        std::cout << "Headphones disconnected! Volume last set at: " << this->volume << std::endl;
-    }
-    unsigned int volume { 3 };
-    bool powerSwitch { true };
-
-    void increaseVolume( unsigned int vol );
-    unsigned int maxThisVolume();
-    unsigned int holdDecreaseVolume( unsigned int vol, unsigned int target );
-    unsigned int holdDecreaseThisVolume ( unsigned int target );
-    void printStatus(); 
-    void maxVol() { std::cout << "Increasing volume to: " << this->maxThisVolume() << std::endl; }
-    JUCE_LEAK_DETECTOR(Headphones)
-};
-
-Headphones::Headphones() {}
-
-void Headphones::increaseVolume ( unsigned int vol )
-{
-    volume = vol;
-    std::cout << "Volume set at: " << vol << std::endl;
-}
-
-unsigned int Headphones::maxThisVolume ( )
-{
-    this->volume = 10;
-    return this->volume;
-    
-}
-
-unsigned int Headphones::holdDecreaseVolume ( unsigned int vol, unsigned int target )
-{
-    if ( vol > target )
-    {
-        while( vol >= target )
-        {
-            vol -= 1;
-            if( vol == target )
-            {
-                volume = vol;
-                return vol;
-            }
-        }
-    }
-
-    volume = 0;
-    return 0;
-}
-
-unsigned int Headphones::holdDecreaseThisVolume ( unsigned int target )
-{
-    if ( this->volume > target )
-    {
-        while( this->volume >= target )
-        {
-            this->volume -= 1;
-            if( this->volume == target )
-            {
-                return this->volume;
-            }
-        }
-    }
-
-    this->volume = 0;
-    return 0;
-}
-
-void Headphones::printStatus()
-{ 
-    std::cout << "Headphone Volume: " << this->volume << ". Power Switch: " << this->powerSwitch<< std::endl;
-
-}
-
-struct HeadphonesWrapper
-{
-    HeadphonesWrapper( Headphones* ptr) : pointerToHeadphones( ptr ) { }
-    ~HeadphonesWrapper()
-    {
-        delete pointerToHeadphones;
-    }
-    
-    Headphones* pointerToHeadphones = nullptr;
-};
-/*
- copied UDT 2:
- */
-struct GiftCard
-{
-    GiftCard();
-    ~GiftCard()
-    {
-        if( this->cardBalance > 5.00 )
-        {
-            std::cout << "Oops! There goes $" << this->cardBalance << "!" << std::endl;
-        }
-        else if ( this->expirationDate < 20200204 )
-        {
-            std::cout << "I really wish I spent that sooner... " << std::endl;
-        }
-    }
-
-    double cardNumber = 6155200244839177.0; //removed const from previous assignment to allow overwrite in constructor
-    unsigned int expirationDate = 20221231; //yyyymmdd
-    double cardBalance = 100.00;
-    
-
-    void makePurchase(double itemCost);
-    void makeThisPurchase(double itemCost);
-    double reloadCard(double addBalance);
-    double reloadThisCard(double addBalance);
-    void recurringPurchase(double subscriptionCost, int subscriptionLength);
-    void thisRecurringPurchase(double subscriptionCost, int subscriptionLength);
-    JUCE_LEAK_DETECTOR(GiftCard)
-};
-
-GiftCard::GiftCard() : cardNumber(2833730365431231.0), expirationDate(20230131), cardBalance(150.00)
-{
-
-}
-void GiftCard::makePurchase(double itemCost)
-{
-    if( expirationDate >= 20200131 )
-    {
-        cardBalance -= itemCost;
-
-        std::cout << "Your remaining balance is: " << cardBalance << std::endl;
-    }
-    else
-    {
-        std::cout << "Card expired!" << std::endl;
-    }
-}
-
-void GiftCard::makeThisPurchase(double itemCost)
-{
-    if( this->expirationDate >= 20200131 )
-    {
-        this->cardBalance -= itemCost;
-
-        std::cout << "Your remaining gift card balance is: " << this->cardBalance << std::endl;
-    }
-    else
-    {
-        std::cout << "Card expired on: " << this->expirationDate << std::endl;
-    }
-}
-
-double GiftCard::reloadCard(double addBalance)
-{
-    cardBalance += addBalance;
-    return cardBalance;
-}
-
-double GiftCard::reloadThisCard(double addBalance)
-{
-    this->cardBalance += addBalance;
-    std::cout << "New gift card balance is " << this->cardBalance << std::endl;
-    return this->cardBalance;
-}
-
-void GiftCard::recurringPurchase(double subscriptionCost, int subscriptionLength)
-{
-    for( int i = 0; i < subscriptionLength; ++i)
-    {
-        cardBalance -= subscriptionCost;
-        std::cout << "Card billed $" << subscriptionCost << ". Your current balance is " << cardBalance << " and your subscription is active for " << subscriptionLength << " more months. " << std::endl;
-    }
-}
-
-void GiftCard::thisRecurringPurchase(double subscriptionCost, int subscriptionLength)
-{
-    for( int i = 0; i < subscriptionLength; ++i)
-    {
-        this->cardBalance -= subscriptionCost;
-        std::cout << "Card billed $" << subscriptionCost << ". Your current balance is " << this->cardBalance << " and your subscription is active for " << subscriptionLength << " more months. " << std::endl;
-    }
-}
-
-struct GiftCardWrapper
-{
-    GiftCardWrapper( GiftCard* ptr) : pointerToGiftCard( ptr ) { }
-    ~GiftCardWrapper()
-    {
-        delete pointerToGiftCard;
-    }
-    
-    GiftCard* pointerToGiftCard = nullptr;
-};
-/*
- copied UDT 3:
- */
-struct Television
-{   
-    Television();
-    ~Television()
-    {   
-        for( int i = 0; i < this->channel; ++i )
-        {
-            if( i % 2 == 0 )
-            {
-                std::cout << "Foo" << std::endl;
-            }
-            else
-            {
-                 std::cout << "Bar" << std::endl;
-            }
-        }
-    }
-    bool isTvOn = false;
-    int volume = 15;
-    int channel = 48;
-    
-    void setVolume( int vol );
-    void setThisVolume( int volume );
-    void changeChannel( int changeTo );
-    void changeThisChannel( int channel );
-    bool powerSwitch( bool power );
-    void playThroughHeadphones( Headphones headphones );
-    void playThisThroughHeadphones( Headphones headphones );
-    void channelUpOrDown( int changeTo );
-    void thisChannelUpOrDown( int changeTo );
-    JUCE_LEAK_DETECTOR(Television)
-};
-
-Television::Television() {}
-
-void Television::setVolume( int vol )
-{
-    volume = vol;
-
-    std::cout << "Volume now set at: " << vol << std::endl;
-
-}
-
-void Television::setThisVolume( int vol )
-{
-    this->volume = vol;
-
-    std::cout << "Volume now set at: " << this->volume << std::endl;
-
-}
-
-void Television::changeChannel( int changeTo )
-{
-    int temp = channel; 
-
-    channel = changeTo; 
-
-    std::cout << "Channel changed from " << temp << " to " << changeTo << std::endl;
-}
-
-void Television::changeThisChannel( int c )
-{
-    int temp = this->channel; 
-
-    this->channel = c; 
-
-    std::cout << "Channel changed from " << temp << " to " << this->channel << std::endl;
-}
-
-bool Television::powerSwitch( bool power )
-{
-    if( power == true )
-    {
-        std::cout << "Powering on!" << std::endl;
-        return true;
-    }
-
-    std::cout << "Shutting down." << std::endl;   
-    return false;
-}
-
-void Television::playThroughHeadphones( Headphones headphones )
-{
-    volume = 0; 
-    headphones.powerSwitch = true;
-}
-
-void Television::playThisThroughHeadphones( Headphones headphones )
-{
-    this->volume = 0; 
-    headphones.powerSwitch = true;
-}
-
-void Television::channelUpOrDown( int changeTo )
-{
-    if( changeTo == channel ) 
-    {
-        std::cout << "Already on channel " << changeTo << std::endl;
-    }
-    else if( changeTo > channel ) 
-    {
-        while( channel != changeTo )
-        {
-            ++channel; 
-            std::cout << "Channel: " << channel << std::endl;
-        }
-    }
-    else // changeTo < channel
-    {
-        int temp = channel - changeTo;
-        for( int i = 0; i < temp; ++i )
-        {
-            --channel;
-            std::cout << "Channel: " << channel << std::endl;
-        }
-    }
-}
-
-
-void Television::thisChannelUpOrDown( int c )
-{
-    if( c == this->channel ) 
-    {
-        std::cout << "Already on channel " << this->channel << std::endl;
-    }
-    else if( c > this->channel ) 
-    {
-        while( c != this->channel )
-        {
-            ++this->channel; 
-            std::cout << "Channel: " << this->channel << std::endl;
-        }
-    }
-    else // changeTo < channel
-    {
-        int temp = this->channel - c;
-        for( int i = 0; i < temp; ++i )
-        {
-            --this->channel;
-            std::cout << "Channel: " << this->channel << std::endl;
-        }
-    }
-}
-
-struct TelevisionWrapper
-{
-    TelevisionWrapper( Television* ptr) : pointerToTelevision( ptr ) { }
-    ~TelevisionWrapper()
-    {
-        delete pointerToTelevision;
-    }
-    
-    Television* pointerToTelevision = nullptr;
-};
-/*
- new UDT 4:
- */
-struct AirPods
-{
-    AirPods();
-    ~AirPods()
-    {
-        pods.maxThisVolume();
-        std::cout << "Air Pods out of charge!" << std::endl;
-    }
-
-    Headphones pods;
-    JUCE_LEAK_DETECTOR(AirPods)
-};
-
-AirPods::AirPods() {}
-
-struct AirPodsWrapper
-{
-    AirPodsWrapper( AirPods* ptr) : pointerToAirPods( ptr ) { }
-    ~AirPodsWrapper()
-    {
-        delete pointerToAirPods;
-    }
-    
-    AirPods* pointerToAirPods = nullptr;
-};
-/*
- new UDT 5:
- */
-struct HomeTheater
-{
-    HomeTheater();
-    ~HomeTheater()
-    {
-        bigScreen.thisChannelUpOrDown(3);
-        std::cout << "Shutting down headphones, volume at: " << turtleBeach.holdDecreaseThisVolume(0) << "." << std::endl;
-    }
-    
-    Television bigScreen;
-    Headphones turtleBeach;
-    JUCE_LEAK_DETECTOR(HomeTheater)
-};
-
-HomeTheater::HomeTheater() {}
-
-struct HomeTheaterWrapper
-{
-    HomeTheaterWrapper( HomeTheater* ptr) : pointerToHomeTheater( ptr ) { }
-    ~HomeTheaterWrapper()
-    {
-        delete pointerToHomeTheater;
-    }
-    
-    HomeTheater* pointerToHomeTheater = nullptr;
-};
-
+#include "AirPods.h"
+#include "GiftCard.h"
+#include "Headphones.h"
+#include "HomeTheater.h"
+#include "Wrappers.h"
 #include <iostream>
+
 int main()
 {
     HeadphonesWrapper beats( new Headphones() );
